requirements.txt
pyspark
pandas
python-dotenv
requests
mysql-connector-python